+++ GrapholizeApp/src/main/java/Controls/TimelineElement/TimeLineElement.java
@@ -8,18 +8,18 @@ import javafx.scene.shape.Rectangle;
 
 public abstract class TimeLineElement extends Rectangle {
 
-    protected long timeStart;
-    protected long timeStop;
+    protected double timeStart;
+    protected double timeStop;
     protected Parent parent;
 
     //Parent not necessary anymore. can be removed from constructor
-    public TimeLineElement(long tStart, long tEnd, double parentHeight, Parent p){
+    public TimeLineElement(double tStart, double tEnd, double parentHeight, Parent p){
         this.timeStart = tStart;
         this.timeStop = tEnd;
         this.parent = p;
 
         setHeight(parentHeight);
-        long width = tEnd - tStart;
+        double width = tEnd - tStart;
         setWidth(width);
     }
 }
